backend/pyproject.toml
# pyproject.toml
[tool.poetry]
name = "ml-backend"
version = "0.1.0"
description = "Modern Python backend with ML capabilities"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.23"
pydantic = {version = "^2.4.2", extras = ["email"]}
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-multipart = "^0.0.6"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
scikit-learn = "^1.3.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
torch = "^2.0.0"
tensorflow = "^2.13.0"
joblib = "^1.3.0"
transformers = "^4.34.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"